Provide cassandra for travis

+++ bin/integration-test.sh
@@ -2,15 +2,21 @@
 
 cli=${1:?Please provide the commandline interface executable as first argument}
 
-set -eu
+set -e
 
 CONTAINER_NAME=db
 CASSANDRA_HOST=127.0.0.1
-CASSANDRA_PORT=12423
+if [ -z "$TRAVIS" ]; then
+    CASSANDRA_PORT=12423
+else
+    CASSANDRA_PORT=9042
+fi
 
 function start_dependencies() {
     echo starting dependencies
-    docker run --rm --name $CONTAINER_NAME -d -p $CASSANDRA_HOST:$CASSANDRA_PORT:9042 --expose 9042 cassandra:2.1 1>&2
+    if [ -z "$TRAVIS" ]; then
+        docker run --rm --name $CONTAINER_NAME -d -p $CASSANDRA_HOST:$CASSANDRA_PORT:9042 --expose 9042 cassandra:2.1 1>&2
+    fi
     while ! nc -d $CASSANDRA_HOST $CASSANDRA_PORT -w 1 -G 1; do
         echo "Waiting for cassandra on $CASSANDRA_HOST:$CASSANDRA_PORT" 1>&2
         sleep 1
@@ -19,8 +25,10 @@ function start_dependencies() {
 }
 
 function stop_dependencies() {
-    echo stopping dependencies ...
-    docker rm --force $CONTAINER_NAME >/dev/null
+    if [ -z "$TRAVIS" ]; then
+        echo stopping dependencies ...
+        docker rm --force $CONTAINER_NAME >/dev/null
+    fi
 }
 
 trap stop_dependencies 0 1 2 5 15
